+++ source_files/edge/w_wad.cc
@@ -631,7 +631,7 @@ static void AddLump(data_file_c *df, int lump, int pos, int size, int file,
 	lump_p->kind = LMKIND_Normal;
 
 	Z_StrNCpy(lump_p->name, name, 8);
-	for (int i=0;i<strlen(lump_p->name);i++) {
+	for (size_t i=0;i<strlen(lump_p->name);i++) {
 		lump_p->name[i] = toupper(lump_p->name[i]);
 	}
  
@@ -1134,7 +1134,7 @@ static void AddFile(const char *filename, int kind, int dyn_index)
                 I_Error("Filename base of %s >8 chars", filename);
 
             strcpy(lump_name, base.c_str());
-			for (int i=0;i<strlen(lump_name);i++) {
+			for (size_t i=0;i<strlen(lump_name);i++) {
 				lump_name[i] = toupper(lump_name[i]);
 			}
         }
